requirements.txt
cassandra-driver==3.28.0
python-dotenv==1.0.0
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
nasdaq-data-link==1.0.4